requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
jinja2>=3.1.0,<4.0.0
python-multipart>=0.0.9,<1.0.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
